+++ apps/masters/serializers.py
@@ -195,3 +195,38 @@ class PurchaseTypesSerializer(serializers.ModelSerializer):
     class Meta:
         model = PurchaseTypes
         fields = '__all__'
+
+class ModGstTypesSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = GstTypes
+        fields = ['gst_type_id','name']
+
+class ModSaleTypesSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = SaleTypes
+        fields = ['sale_type_id','name']
+
+class SaleTypesSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = SaleTypes
+        fields = '__all__'
+
+class ShippingModesSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = ShippingModes
+        fields = '__all__'
+
+class GstTypesSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = GstTypes
+        fields = '__all__'
+
+class ModShippingCompaniesSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = ShippingCompanies
+        fields = ['shipping_company_id','code','name']
+
+class ShippingCompaniesSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = ShippingCompanies
+        fields = '__all__'
