+++ apps/masters/urls.py
@@ -25,6 +25,10 @@ router.register(r'product-item-type', ProductItemTypeViewSet)
 router.register(r'brand-salesman', BrandSalesmanViewSet)
 router.register(r'product-brands', ProductBrandsViewSet)
 router.register(r'purchase-types', PurchaseTypesViewSet)
+router.register(r'shipping_companies', ShippingCompaniesView)
+router.register(r'shipping_modes', ShippingModesView)
+router.register(r'sale_types', SaleTypesView)
+router.register(r'gst_types', GstTypesView)
 
 urlpatterns = [
     path('', include(router.urls)),
