+++ apps/sales/serializers.py
@@ -1,51 +1,22 @@
 from rest_framework import serializers
 from apps.customer.serializers import ModCustomerAddressesSerializer, ModCustomersSerializer, ModCustomerPaymentTermsSerializers, ModLedgerAccountsSerializers
-from apps.masters.serializers import ModCustomerCategoriesSerializers, ModProductBrandsSerializer
+from apps.inventory.serializers import ModWarehousesSerializer
+from apps.masters.serializers import ModCustomerCategoriesSerializers, ModGstTypesSerializer, ModProductBrandsSerializer, ModSaleTypesSerializer, ModShippingCompaniesSerializer, ShippingModesSerializer
 from apps.products.serializers import ModProductGroupsSerializer, ModproductsSerializer
 from .models import *
 
 
 #----------Modified Serializers--------------------------
-class ModGstTypesSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = GstTypes
-        fields = ['gst_type_id','name']
-
-class ModSaleTypesSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = SaleTypes
-        fields = ['sale_type_id','name']
 
 class ModSaleOrderSerializer(serializers.ModelSerializer):
     class Meta:
         model = SaleOrder
         fields = ['order_id','customer_id','order_date','delivery_date']
 
-class ModShippingCompaniesSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = ShippingCompanies
-        fields = ['shipping_company_id','code','name']
 # -------------------------------------------------------
 
-class SaleTypesSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = SaleTypes
-        fields = '__all__'
 
-class ShippingModesSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = ShippingModes
-        fields = '__all__'
 
-class GstTypesSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = GstTypes
-        fields = '__all__'
-
-class ShippingCompaniesSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = ShippingCompanies
-        fields = '__all__'
 
 class SaleOrderSerializer(serializers.ModelSerializer):
     gst_type = ModGstTypesSerializer(source='gst_type_id', read_only=True)
@@ -61,7 +32,7 @@ class SaleOrderSerializer(serializers.ModelSerializer):
 
 class InvoicesSerializer(serializers.ModelSerializer):
     order = ModSaleOrderSerializer(source='order_id', read_only=True)
-    # warehouse_id = ModSaleOrderSerializers(source='order_id', read_only=True) #update here later
+    warehouse = ModWarehousesSerializer(source='warehouse_id', read_only=True)
     sale_type = ModSaleTypesSerializer(source='sale_type_id', read_only=True)
 
     class Meta:
@@ -69,7 +40,7 @@ class InvoicesSerializer(serializers.ModelSerializer):
         fields = '__all__'
 
 class PaymentTransactionsSerializer(serializers.ModelSerializer):
-    invoice = InvoicesSerializer(source='invoice_id', read_only=True) #change the serializers
+    invoice = InvoicesSerializer(source='invoice_id', read_only=True)
     
     class Meta:
         model = PaymentTransactions
