+++ apps/sales/urls.py
@@ -5,17 +5,14 @@ from rest_framework.routers import DefaultRouter
 #add your urls
 
 router = DefaultRouter()
-router.register(r'sale-types', SaleTypesView)
-router.register(r'sale-order', SaleOrderView)
-router.register(r'shipping-modes', ShippingModesView)
-router.register(r'gst-types', GstTypesView)
-router.register(r'shipping-companies', ShippingCompaniesView)
-router.register(r'payment-transactions', PaymentTransactionsView)
+
+router.register(r'sale_order', SaleOrderView)
+router.register(r'payment_transactions', PaymentTransactionsView)
 router.register(r'shipments', ShipmentsView)
-router.register(r'order-items', OrderItemsView)
+router.register(r'order_items', OrderItemsView)
 router.register(r'invoices', InvoicesView)
-router.register(r'sales-price-list', SalesPriceListView)
-router.register(r'sale-order-returns', SaleOrderReturnsView)
+router.register(r'sales_price_list', SalesPriceListView)
+router.register(r'sale_order_returns', SaleOrderReturnsView)
 
 urlpatterns = [
     path('',include(router.urls))
